+++ tests/specs/pat/image-crop.js
@@ -1,6 +1,6 @@
 define(["pat-image-crop"], function(pattern) {
 
-    describe("image-crop-pattern", function() {
+    describe("pat-image-crop", function() {
 
         beforeEach(function() {
             $("<div/>", {id: "lab"}).appendTo(document.body);